Add files via upload
This is the dll implementation of the Codeplex project Tailf (http://tailf.codeplex.com/) of Felix Pollan with some enhancements.
The dll implementation is same as in the tail.cs of the tailf project, but just added possibility to get lines in list format if there is no event handler. The dll project is added to the original Visual Studio solution.